adds country app
