+++ server/country/admin.py
@@ -1,4 +1,5 @@
 from django.contrib import admin
+from .models import Country
 
 # Register your models here.
-
+admin.site.register(Country)
