+++ server/country/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from country.views import get_all_countries
+
+urlpatterns = [
+    path('/all', get_all_countries, name='get_all_countries'),
+]
