--- server/world_proxy/serializers.py
@@ -1,10 +1,0 @@
-from rest_framework import serializers
-from .models import Country
-
-
-class CountrySerializer(serializers.ModelSerializer):
-    expected = serializers.Field(source='actual')
-
-    class Meta:
-        model = Country
-        fields = ('field1', 'field2', 'expected')
