+++ server/world_proxy/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
-from world_proxy.views import welcome, get_all_countries
+from world_proxy.views import welcome
+
 
 urlpatterns = [
     path('', welcome, name='welcome'),
-    path('countries/all', get_all_countries, name='get_all_countries')
 ]
